updated controller with Rowan's suggestions
